ux improvements + code refactoring
