+++ src/routes/status.tsx
@@ -5,286 +5,19 @@
  * found in the LICENSE file in the root directory of this source tree.
  */
 
-import { Badge, Button, Spin, Table, List } from 'antd';
-import { ColumnProps } from 'antd/lib/table';
-import { Link } from 'react-router-dom';
-import BN from 'bn.js';
-import { Input, Outpoint, Tx } from 'leap-core';
-import { action, observable, reaction, runInAction } from 'mobx';
 import { inject, observer } from 'mobx-react';
 import * as React from 'react';
-import { EventLog } from 'web3/types';
 import AppLayout from '../components/appLayout';
 import { default as Monitor } from '../components/monitor';
-import TokenValue from '../components/tokenValue';
 import { CONFIG } from '../config';
-import Bridge from '../stores/bridge';
-import ExitHandler from '../stores/exitHandler';
-import Explorer from '../stores/explorer';
-import Tokens from '../stores/tokens';
-import Web3Store from '../stores/web3';
-import HexString from '../components/hexString';
-
-interface StatusProps {
-  exitHandler: ExitHandler;
-  explorer: Explorer;
-  bridge: Bridge;
-  web3: Web3Store;
-  tokens: Tokens;
-}
-
-interface ColorDetails {
-  color: number;
-  plasmaBalance: BN;
-  utxoSum: BN;
-  exitSum: BN;
-}
-
-interface ExitData {
-  event: EventLog;
-  exit: {
-    amount: string;
-    color: string;
-    owner: string;
-    finalized: boolean;
-    priorityTimestamp: string;
-    stake: string;
-    0: string;
-    1: string;
-    2: string;
-    3: boolean;
-    4: string;
-    5: string;
-  };
-  suspect: boolean;
-}
-
-@inject('exitHandler', 'explorer', 'web3', 'tokens', 'bridge')
-@observer
-export default class Status extends React.Component<StatusProps, {}> {
-  private renderExit(record) {
-    return (
-      <List.Item>
-        <List.Item.Meta
-          title={
-            <div
-              style={{
-                color: record.suspect ? 'red' : 'green',
-              }}
-            >
-              <TokenValue
-                value={record.exit.amount}
-                color={record.exit.color}
-              />
-            </div>
-          }
-          description={
-            <div>
-              <div>
-                <Link to={`/explorer/address/${record.exit.owner}`}>
-                  <HexString>{record.exit.owner}</HexString>
-                </Link>
-              </div>
-              <div>
-                {new Date(
-                  parseInt(record.exit.priorityTimestamp) * 1000
-                ).toUTCString()}
-              </div>
-            </div>
-          }
-        />
-      </List.Item>
-    );
-  }
-
-  private colorColums: ColumnProps<any>[] = [
-    {
-      title: 'Color',
-      dataIndex: 'color',
-      key: 'color',
-    },
-    {
-      title: 'Plasma balance',
-      dataIndex: 'plasmaBalance',
-      key: 'plasmaBalance',
-      render: (value, record) => (
-        <TokenValue value={value} color={record.color} />
-      ),
-    },
-    {
-      title: 'Utxo Sum',
-      dataIndex: 'utxoSum',
-      key: 'utxoSum',
-      render: (value, record) => (
-        <TokenValue value={value} color={record.color} />
-      ),
-    },
-    {
-      title: 'Open Exit Sum',
-      dataIndex: 'exitSum',
-      key: 'exitSum',
-      render: (value, record) => (
-        <TokenValue value={value} color={record.color} />
-      ),
-    },
-    {
-      title: 'Status',
-      key: 'status',
-      render: (value, record: ColorDetails) => {
-        return record.plasmaBalance.gte(record.exitSum.add(record.utxoSum)) ? (
-          <Badge status="success" text="Ok" />
-        ) : (
-          <Badge status="warning" text="Invalid" />
-        );
-      },
-    },
-  ];
-
-  constructor(props) {
-    super(props);
-    reaction(() => this.props.exitHandler.address, this.getData, {
-      fireImmediately: true,
-    });
-  }
-
-  @observable exitDetails: ExitData[] = [];
-  @observable colorDetails: ColorDetails[] = [];
-
-  @action
-  private getData = async (): Promise<void> => {
-    const { exitHandler, explorer, web3, tokens, bridge } = this.props;
-    const exits = [];
-    if (!bridge.address)
-      await new Promise(resolve => reaction(() => bridge.address, resolve));
-    const fromBlock = await bridge.contract.methods.genesisBlockNumber().call();
-    const events = await exitHandler.contract.getPastEvents('ExitStarted', {
-      fromBlock,
-    });
-    await Promise.all(
-      events.map(async event => {
-        const utxoId = new Outpoint(
-          event.returnValues.txHash,
-          parseInt(event.returnValues.outIndex)
-        ).getUtxoId();
-        const exit = await exitHandler.contract.methods
-          .exits(web3.root.instance.utils.toHex(utxoId))
-          .call();
-        let suspect;
-        const exitTxHash = Tx.exit(
-          new Input(
-            new Outpoint(
-              event.returnValues.txHash,
-              parseInt(event.returnValues.outIndex)
-            )
-          )
-        ).hash();
-        const exitTx = await explorer.getTransaction(exitTxHash);
-        if (exitTx === undefined) {
-          suspect = true;
-        }
-        exits.push({ event, exit, suspect });
-      })
-    );
-
-    const uTxos = await web3.plasma.instance.getUnspentAll();
-    const addresses = await web3.plasma.instance.getColors();
-    const colors = await Promise.all(
-      addresses.map(async addr => {
-        return web3.plasma.instance.getColor(addr);
-      })
-    );
-    const colorDetails = await Promise.all(
-      colors.map(async color => {
-        const utxoSum = uTxos.reduce((acc, utXo) => {
-          return utXo.output.color === color
-            ? web3.root.instance.utils.toBN(utXo.output.value).add(acc)
-            : acc;
-        }, web3.root.instance.utils.toBN(0));
-
-        const exitSum = exits.reduce((acc, e) => {
-          return parseInt(e.exit.color) === color && !e.exit.finalized
-            ? web3.root.instance.utils.toBN(e.exit.amount).add(acc)
-            : acc;
-        }, web3.root.instance.utils.toBN(0));
-
-        const plasmaBalance = web3.root.instance.utils.toBN(
-          await tokens.list
-            .find(token => token.color === color)
-            .contract.methods.balanceOf(exitHandler.address)
-            .call()
-        );
-        return {
-          utxoSum,
-          exitSum,
-          plasmaBalance,
-          color,
-        };
-      })
-    );
-    runInAction(() => {
-      this.exitDetails = exits;
-      this.colorDetails = colorDetails;
-    });
-  };
-
-  private exitDetailsRenderer = record => {
-    if (!this.exitDetails.length) return 'No open exits';
-    const openExits = this.exitDetails.filter(
-      e => !e.exit.finalized && parseInt(e.exit.color) === record.color
-    );
-    console.log('openexits', openExits);
-    const badFinalizedExits = this.exitDetails.filter(
-      e =>
-        parseFloat(e.exit.color) === record.color &&
-        e.exit.finalized &&
-        e.suspect
-    );
-    return (
-      <>
-        <h3>Open exits</h3>
-        <List
-          split
-          pagination={{ pageSize: 10 }}
-          dataSource={openExits}
-          renderItem={this.renderExit}
-        />
-        <h3>Bad finalized exits</h3>
-        <List
-          split
-          pagination={{ pageSize: 10 }}
-          dataSource={badFinalizedExits}
-          renderItem={this.renderExit}
-        />
-      </>
-    );
-  };
-
-  expandRenderer = props => {
-    let text;
-    return (
-      <Button onClick={e => props.onExpand(props.record, e)}>
-        {props.expanded ? 'Hide exits' : 'Show exits'}
-      </Button>
-    );
-  };
+import Watchtower from '../components/watchtower';
+export default class Status extends React.Component<{}, {}> {
   render() {
     return (
       <AppLayout section="status">
         <h1>Nodes status</h1>
         <Monitor nodes={CONFIG.nodes} />
-        {this.colorDetails.length ? (
-          <Table
-            style={{ paddingTop: '2rem' }}
-            bordered
-            dataSource={this.colorDetails}
-            columns={this.colorColums}
-            expandedRowRender={this.exitDetailsRenderer}
-            expandIcon={this.expandRenderer}
-            title={() => <h1>Watchtower</h1>}
-          />
-        ) : (
-          <Spin style={{ margin: 'auto' }} />
-        )}
+        <Watchtower/>
       </AppLayout>
     );
   }
